+++ databases/device.database.js
@@ -16,8 +16,33 @@ function loadDatabase(dbName)
 	createDatabaseFolder(dbFolderPath);
 	loadedDatabaseObject = level(dbFolderPath);
 	
-	loadRes["listDevices"] = placeholder;
-	loadRes["listAllDevices"] = placeholder;
+	
+	function callListDevices(listCallback)
+	{
+		callListAllDevices(false, listCallback);
+	}
+	
+	
+	function callListAllDevices(deleteStatus, listAllCallback)
+	{
+		var retrievedEntries = [];
+		var readStreamObject = loadedDatabaseObject.createReadStream();
+		
+		readStreamObject.on("data", function (currentEntry)
+		{
+			addRetrievedEntry(currentEntry);
+		});
+		
+		readStreamObject.on("end", function()
+		{
+			return listAllCallback(null, retrievedEntries);
+		});
+	}
+	
+	
+	
+	loadRes["listDevices"] = callListDevices;
+	loadRes["listAllDevices"] = callListAllDevices;
 	loadRes["createDeviceEntity"] = placeholder;
 	loadRes["readDeviceEntity"] = placeholder;
 	loadRes["updateDeviceEntity"] = placeholder;
@@ -36,6 +61,26 @@ function createDatabaseFolder(targetPath)
 	if (folderExists !== true)
 	{
 		fs.mkdirSync(targetPath, folderOpts);
+	}
+}
+
+
+function addRetrievedEntry(dataObj, delStat, entryArr)
+{
+	var parsedEntry = {};
+	
+	try
+	{
+		parsedEntry = JSON.parse(dataObj.value);
+		
+		if (delStat === true || parsedEntry.isDeleted !== true)
+		{
+			entryArr.push(parsedEntry);
+		}
+	}
+	catch(e)
+	{
+		console.log(e.message);
 	}
 }