+++ changelog.md
@@ -1,13 +1,30 @@
 # Changelog
 
+**Test Status**
+* G - Controller Files
+	* Emulated Remote IO initialization and reading from database.
+	* Tests passed: 46 / 123.
+
+---
+
 **./databases/device.database.js**
-* New file - Library for interfacing with local databases.
-* This will include full CRUD functionality.
-* So far, it only creates the relevant folders.
-* The actual functions are only placeholders.
+* Wrote new function 'addRetrievedEntry'
+	* When a database entry is retrieved, parse it to JSON and add to result set.
+* Wrote new CRUD functions:
+	* callListDevices
+	* callListAllDevices
 
 ---
 
 **./fox-devices/remote_io/remote-io.index**
-* Added file requirements.
-* Functions have not been defined yet.
+* Declared global variables:
+	* initializationCallbacks
+	* initializationComplete
+	* runningIoDevices
+* Wrote new functions:
+	* 'initializeRemoteIoFactory' - Initializes Remote IO devices.
+	* 'whenInitializationComplete' - Saves initialization callback.
+	* 'crudListRemoteIoDevices' - Retrieves Remote IO devices from database.
+* Wrote new sub functions:
+	* 'runInitializationLoop' - Creates and saves 'StoredDevice' objects from database.
+	* 'runCallbackLoop' - Runs initialization callbacks upon completion.